mcp-project/agents/python-processor/requirements.txt
nats-py==2.3.1
asyncio-nats-client==2.0.0
pydantic==2.0.3
loguru==0.7.0
pytest==7.4.0
pytest-asyncio==0.21.1